webstorage/storage_local_security.sub.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Web Storage</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  document.domain = document.domain;
  assert_throws("SecurityError", function() {
    localStorage
  });
});
</script>
